Issue11: Using Flux and Mono in a proper way (AFAIK for now) in the V2 of rest-api
