+++ api-server/src/main/java/org/opennms/devjam2022/apiserver/listener/ContextListener.java
@@ -1,6 +1,7 @@
 package org.opennms.devjam2022.apiserver.listener;
 
 import java.util.Random;
+import org.opennms.devjam2022.apiserver.impl.InMemoryUserServiceNB;
 import static org.opennms.devjam2022.apiserver.model.utils.ModelUtil.createTestRole;
 import static org.opennms.devjam2022.apiserver.model.utils.ModelUtil.createTestUserWithEmptyListRoles;
 import org.opennms.devjam2022.apiserver.service.IUserService;
@@ -21,15 +22,21 @@ public class ContextListener {
     @Autowired
     IUserService userService;
 
+    @Qualifier("InMemoryUserServiceNB")
+    @Autowired
+    InMemoryUserServiceNB userServiceNb;
+
     @EventListener({ContextRefreshedEvent.class})
     public void handleContextRefreshEvent() {
         LOG.info("Context refresh event received");
         for (long i = 0; i < NUMBER_OF_USERS_IN_DB; i++) {
             String userId = userService.addUser(createTestUserWithEmptyListRoles());
+            String userIdNb = userServiceNb.addUser(createTestUserWithEmptyListRoles());
 
             int count = new Random().nextInt(5) + 1; // 1-5 roles to be added
             for (int j = 0; j < count; j++) {
                 userService.addRole(userId, createTestRole());
+                userServiceNb.addRole(userIdNb, createTestRole());
             }
         }
         LOG.info("Added '" + NUMBER_OF_USERS_IN_DB + "' users... to the in-memory database");
